requirements.txt
pandas>=2.0
numpy>=1.26
openai>=1.0
anthropic>=0.40
mistralai>=1.0
folktables>=0.0.12
scikit-learn>=1.4
sentence-transformers>=3.0
torch>=2.0
tqdm>=4.66
tenacity>=8.0
python-dotenv>=1.0
matplotlib>=3.8
seaborn>=0.13
jupyter>=1.0
